moved app-config to its own file
